D8CORE-5036: Generalize iframe embed validation
