+++ src/Plugin/EmbedValidator/AirtableValidator.php
@@ -2,7 +2,7 @@
 
 namespace Drupal\stanford_media\Plugin\EmbedValidator;
 
-use Drupal\stanford_media\Plugin\EmbedValidatorBase;
+use Drupal\stanford_media\Plugin\AbstractIframeValidator;
 
 /**
  * Airtable Iframe validation.
@@ -12,27 +12,8 @@ use Drupal\stanford_media\Plugin\EmbedValidatorBase;
  *   label = "Airtable"
  * )
  */
-class AirtableValidator extends EmbedValidatorBase {
+class AirtableValidator extends AbstractIframeValidator {
 
-  /**
-   * {@inheritDoc}
-   */
-  public function isEmbedCodeAllowed(string $code): bool {
-    $code = str_replace("\n", ' ', $code);
-    preg_match('/<iframe.* src="(.+?)"/', $code, $matches);
-    if (empty($matches[1])) {
-      return FALSE;
-    }
-    $source = parse_url($matches[1]);
-    return $source['host'] == 'airtable.com';
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  public function prepareEmbedCode(string $code): string {
-    preg_match('/<iframe.*?>/', $code, $modified_code);
-    return $modified_code ? $modified_code[0] . '</iframe>' : '';
-  }
+  const EMBEDDOMAIN = 'airtable.com';
 
 }
